pyproject.toml
[tool.poetry]
name = "gemini-ia-studio"
version = "0.1.0"
description = ""
authors = ["Rodrigo Silva <rodrigo.sousa.06@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.10.1"
google-generativeai = "^0.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
